pyproject.toml
[build-system]
requires = ["setuptools>=70", "setuptools-scm"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "company-doc-finder"
version = "2.0.0"
description = "Self-hosted multi-tenant company document finder & summarizer"
requires-python = ">=3.11"
dependencies = [
    # ── LlamaIndex ────────────────────────────────────────────────
    "llama-index>=0.12",
    "llama-index-embeddings-huggingface>=0.5",
    "llama-index-llms-ollama>=0.5",
    "llama-index-vector-stores-qdrant>=0.6",
    "llama-index-node-parser-semantic>=0.1",
    # LlamaCloud (optional — only needed for cloud providers)
    "llama-parse>=0.5",

    # ── API ───────────────────────────────────────────────────────
    "fastapi[standard]>=0.115",
    "pydantic-settings>=2.3",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",

    # ── Task queue — VALKEY (not redis) ───────────────────────────
    "celery[valkey]>=5.4",
    "valkey>=6.0",

    # ── Database — MongoDB ────────────────────────────────────────
    "motor>=3.4",
    "beanie>=1.26",
    "pymongo>=4.8",

    # ── Document parsing ──────────────────────────────────────────
    "unstructured[pdf,docx]>=0.15",
    "python-magic>=0.4",

    # ── Ingestion utilities ───────────────────────────────────────
    "watchdog>=4.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "httpx>=0.27",
    "tenacity>=8.3",

    # ── LangChain (for RecursiveCharacterTextSplitter + LangchainNodeParser) ──
    "langchain-text-splitters>=0.3",
    "langchain-ollama>=0.2",
    "langchain-huggingface>=0.1",

    # ── Evaluation ────────────────────────────────────────────────
    "ragas>=0.2",

    # ── Observability ─────────────────────────────────────────────
    "structlog>=25.0",
    "openinference-instrumentation-llama-index>=3.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "opentelemetry-sdk>=1.25",
    "prometheus-fastapi-instrumentator>=7.0",

    # ── Rate limiting ─────────────────────────────────────────────
    "slowapi>=0.1",

    # ── Security & Scanning ───────────────────────────────────────
    "pyclamd>=0.4",

    # ── Backup - S3/MinIO ─────────────────────────────────────────
    "boto3>=1.34",
    "s3fs>=2024.6",

    # ── GPU Embedding Worker (optional) ───────────────────────────
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.setuptools.package-dir]
"" = "."

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
